101-quote.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %d byte(s)\n", (long unsigned int)sizeof(char));
-	printf("Size of an int: %d byte(s)\n" (long unsigned int)sizeof(int));
-	printf("Size of a long int: %d bytes(s) \n", (long unsigned int)sizeof(long int));
-	printf("Size of a long long int: %ld byte(s)\n", (long unsigned int)sizeof(long long int));
-	printf("Size of float: %d byte(s)\n", (unsigned long int)sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int)sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int)sizeof(int));
+	printf("Size of a long int: %d bytes(s) \n", (int)sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (int)sizeof(long long int));
+	printf("Size of float: %d byte(s)\n", (int)sizeof(float));
 	return (0);
 }
